U teuton 2.1.8 and may other improvements

+++ src/main/java/io/github/teuton/panel/ui/components/CaseListCell.java
@@ -54,7 +54,7 @@ public class CaseListCell extends JFXListCell<Case> {
 			}
 			icon.setIconColor(gradeColor);
 
-			setText("Case " + ((Double)c.getResults().get("case_id")).intValue());
+			setText("Case " + c.getResults().get("case_id"));
 			setGraphic(icon);
 
 		}
